requirements.txt
ffmpeg-python == 0.2.0
m3u8 == 0.5.2
requests == 2.22.0